+++ packages/kofono/src/extension/UpdateCounter/UpdateCounterExtension.ts
@@ -1,6 +1,6 @@
 import { Events } from "../../form/events/types";
-import type { Form } from "../../form/Form";
 import { BaseExtension } from "../BaseExtension";
+import type { ExtensionContext, SchemaExtensionBaseOptions } from "../types";
 
 // this is a basic extension to count the number of updates performed on the form
 // it serves as an example of how to create a custom extension
@@ -11,7 +11,7 @@ export type SchemaUpdateCounterExtension = {
 };
 
 // represent the options passed to the extension at form creation.
-export type UpdateCounterOpts = unknown;
+export type UpdateCounterOpts = SchemaExtensionBaseOptions;
 
 // represent the extension meta-state shape/type
 export type UpdateCounterMeta = number;
@@ -19,31 +19,33 @@ export type UpdateCounterMeta = number;
 // represent the extension declaration and factory
 export const updateCounterExtension = {
     name: "updateCounter" as const,
-    factory: (opts: UpdateCounterOpts) => new UpdateCounterExtension(opts),
+    factory: (ctx: ExtensionContext, opts: UpdateCounterOpts) =>
+        new UpdateCounterExtension(ctx, opts),
 };
 
-// function to build the schema definition
-export function updateCounter(): SchemaUpdateCounterExtension {
+// helper to build the schema definition
+export function updateCounter(id?: string): SchemaUpdateCounterExtension {
     return {
-        updateCounter: {},
+        updateCounter: {
+            id,
+        },
     };
 }
 
 // the extension implementation
-export class UpdateCounterExtension extends BaseExtension<UpdateCounterMeta> {
-    public readonly metaName: string = updateCounterExtension.name;
-    public readonly metaData = 0;
+export class UpdateCounterExtension extends BaseExtension<
+    UpdateCounterMeta,
+    UpdateCounterOpts
+> {
+    public readonly defaultMetaData: UpdateCounterMeta = 0;
 
-    constructor(_: UpdateCounterOpts = {}) {
-        super();
+    constructor(ctx: ExtensionContext, opts: UpdateCounterOpts = {}) {
+        super(ctx, opts);
     }
 
-    async init(form: Form): Promise<void> {
-        this.initMeta(form);
-        let _metaValue = this.meta(form);
-
-        form.events.on(Events.SelectorAfterUpdate, () => {
-            _metaValue++;
+    async init(): Promise<void> {
+        this.form.events.on(Events.SelectorAfterUpdate, () => {
+            this.metaData = this.metaData + 1;
         });
     }
 }